Fixed the ending time issue
Leon was saying if he makes an event that begins on 3/1/2015 at 103000 and have it end on 3/21/2015 at 083000, program doesn't work. Modified the if statement that compares starting and ending time to also check the dates. If user specified a valid ending date, the ending time will not get rejected.